+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Left2.java
@@ -6,6 +6,7 @@ import com.qualcomm.hardware.bosch.JustLoggingAccelerationIntegrator;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
+import com.qualcomm.hardware.rev.RevColorSensorV3;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
@@ -13,6 +14,8 @@ import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
 import com.qualcomm.robotcore.hardware.CRServo;
+import com.qualcomm.robotcore.hardware.Servo;
+
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;
@@ -55,7 +58,11 @@ public class Left2 extends LinearOpMode {
     DcMotor Spin;
     DcMotor Crane;
     CRServo Left;
+
+    Servo Guide;
+
     Rev2mDistanceSensor distance;
+    Rev2mDistanceSensor distance2;
 
 
     BNO055IMU imu;
@@ -167,10 +174,12 @@ public class Left2 extends LinearOpMode {
         backLeft = hardwareMap.get(DcMotor.class, "backLeft");                      //mapping the motors
         backRight = hardwareMap.get(DcMotor.class, "backRight");
 
+        Guide = hardwareMap.get(Servo.class, "Guide");
         Left = hardwareMap.get(CRServo.class, "Lefts");
         Spin = hardwareMap.get(DcMotor.class, "Spin");
         Crane = hardwareMap.get(DcMotor.class, "Crane");
 
+        distance2=hardwareMap.get(Rev2mDistanceSensor.class,"distance 2");
         distance=hardwareMap.get(Rev2mDistanceSensor.class,"distance");
 
         frontLeft.setDirection(DcMotorSimple.Direction.FORWARD);
@@ -178,67 +187,107 @@ public class Left2 extends LinearOpMode {
         backLeft.setDirection(DcMotorSimple.Direction.FORWARD);
         backRight.setDirection(DcMotorSimple.Direction.FORWARD);
 
-
         if (opModeIsActive()) {//start of queue for autonnums movment
+            Spin.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
             Left.setPower(.3);
-            sleep(500);
-            crane(-1,400);
-            strafeLeftwithcrane2(1,30,-1,2000);
-            //strafeLeftwithcrane(1,1950,-1,2300);
-            //crane2(1,-7600);
-            move(.5,330);
-            sleep(1000);
-            move(.5,40);
-            sleep(200);
-            crane(1,200);
-            intake(-1,1300);
-            crane(-1,250);
-            move(.5,-200);
-            //new
-           /*
-            strafeLeftwithcrane(1,600,1,1000);//500
-            gyroTurning(0);
-            sleep(1000);
-            moveandspin(.5,-980,-1,1000);///410
-            stopMotors();
-            //spin2(-1,1000);
-            move(.2,-150);
-            craneinput(500);
-            Left.setPower(.3);
-            crane(-1, 650);
-            gyroTurning(0);
-            moveandspin(.8,1000,1,1000);
-            strafeRightwithcrane(1,680,-1,1900);
-            move(.5,360);
-            intake(-1,1300);
-            move(.5,-330);
-*/
 
-            switch (location){//determine where to park
-                case 0:
-                    //move(.2,-200);
-                    stopMotors();
-                    sleep(3000);
-                    break;
-                case 1:
-                    strafeRight(1,700);
-                    gyroTurning(0);
-                    move(.4,-1300);
-                    //strafeLeft(.5,300);
-                    sleep(3000);
-                    break;
-                case 2:
-                    //move(.2,-200);
-                    stopMotors();
-                    sleep(3000);
-                    break;
-                case 3:
-                    strafeRight(1,700);
-                    gyroTurning(0);
-                    move(.6,1000);
-                    //strafeLeft(.5,300);
-                    sleep(3000);
-                    break;
+            //move and spin to high junction
+            strafeLeftwithcrane(1, 2700, -1, 3500);
+            spin(1,260);
+            Crane.setPower(-.8);
+            gyroTurning(0);
+            sleep(100);
+
+            //drop cone
+            move(.5,-450);
+            Guide.setPosition(0);
+            sleep(1000);
+            move(.5,-100);
+            stopMotors();
+            sleep(1000);
+            crane(1,100);
+            Guide.setPosition(1);
+            intake(-1, 500);
+
+            //strafe off junction
+            crane(-1,550);
+            strafeRightwithcrane(1, 350, 1, 650);
+
+            //move to stack
+            stopMotors();
+            gyroTurning(0);
+            moveandspin(1, -1200, 1, 591);
+            sleep(100);
+            gyroTurning(0);
+            movewithdistance(.4, 1);
+            //pick up cone
+            craneinput(600);
+            crane(-1, 500);
+            Left.setPower(.2);
+
+            //move to high junction
+            move(.4, 400);
+            gyroTurning(0);
+            moveandspinandcrane(.8, 950, 1, 260, -1, 2600);//-315
+            Crane.setPower(-.8);
+            Guide.setPosition(0);
+            sleep(1000);
+            movewithdistance2(.5, 12);
+            move(.5, -130);
+            stopMotors();
+            //dropcone
+            crane(1,100);
+            Guide.setPosition(1);
+            intake(-1, 500);
+
+            //strafe off
+            Crane.setPower(-.6);
+            strafeRightwithcrane(.2, 200, 1, 1000);
+            gyroTurning(0);
+
+            if (getRuntime() < 22) {
+                //////////////////////////////////////////////////////////////////not made for side past this
+                //move back to stack
+                moveandspin(.8, 1100, 1, -591);
+                move(.2, 400);
+                //pick up another cone
+                craneinput(400);
+                crane(-1, 1000);
+                //move to large junction again
+                moveandspinandcrane(.8, -1500, 1, -315, -1, 1500);
+                Crane.setPower(-.2);
+                stopMotors();
+                strafeLeft(.2, 200);
+                //release cone
+                intake(-1, 1000);
+                strafeRightwithcrane(.2, 200, 1, 1000);
+
+
+                switch (location) {//determine where to park
+                    case 0:
+                        move(.2, -300);
+                        stopMotors();
+                        sleep(3000);
+                        break;
+                    case 1:
+
+                        gyroTurning(0);
+                        move(.4, 800);
+                        stopMotors();
+                        sleep(3000);
+                        break;
+                    case 2:
+                        move(.2, -300);
+                        stopMotors();
+                        sleep(3000);
+                        break;
+                    case 3:
+                        gyroTurning(0);
+                        move(.4, -2500);
+                        stopMotors();
+                        sleep(3000);
+                        break;
+                }
             }
         }
     }
@@ -278,41 +327,41 @@ public class Left2 extends LinearOpMode {
             telemetry.addData("Angle", currentAngle);
             telemetry.addData("targetangle", targetAngle);
             telemetry.update();
-            if (angles.firstAngle >= targetAngle - 0.1 && angles.firstAngle <= targetAngle + 0.1) {
+            if (angles.firstAngle >= targetAngle - 0.15 && angles.firstAngle <= targetAngle + 0.15) {
                 frontLeft.setPower(0);
                 frontRight.setPower(0);
                 backLeft.setPower(0);
                 backRight.setPower(0);
                 foundAngle = true;
-                sleep(1000);
+                sleep(500);
                 break;
 
             } else if (angles.firstAngle >= targetAngle + 0.5) {
                 if (angles.firstAngle <= targetAngle - 5) {
-                    frontLeft.setPower(0.25);
-                    frontRight.setPower(-0.25);
-                    backLeft.setPower(0.25);
-                    backRight.setPower(-0.25);
+                    frontLeft.setPower(0.3);
+                    frontRight.setPower(-0.3);
+                    backLeft.setPower(0.3);
+                    backRight.setPower(-0.3);
                     foundAngle = false;
                 } else {
-                    frontLeft.setPower(-0.25);
-                    frontRight.setPower(0.25);
-                    backLeft.setPower(-0.25);
-                    backRight.setPower(0.25);
+                    frontLeft.setPower(-0.3);
+                    frontRight.setPower(0.3);
+                    backLeft.setPower(-0.3);
+                    backRight.setPower(0.3);
                     foundAngle = false;
                 }
             } else if (angles.firstAngle <= targetAngle - 0.5) {
                 if (angles.firstAngle >= targetAngle + 5) {
-                    frontLeft.setPower(-0.25);
-                    frontRight.setPower(0.25);
-                    backLeft.setPower(-0.25);
-                    backRight.setPower(0.25);
+                    frontLeft.setPower(-0.3);
+                    frontRight.setPower(0.3);
+                    backLeft.setPower(-0.3);
+                    backRight.setPower(0.3);
                     foundAngle = false;
                 } else {
-                    frontLeft.setPower(.25);
-                    frontRight.setPower(-.25);
-                    backLeft.setPower(.25);
-                    backRight.setPower(-.25);
+                    frontLeft.setPower(.3);
+                    frontRight.setPower(-.3);
+                    backLeft.setPower(.3);
+                    backRight.setPower(-.3);
                     foundAngle = false;
                 }
             }
@@ -360,6 +409,11 @@ public class Left2 extends LinearOpMode {
 
     }
     public void moveandspin(double power, int position,double powers, int times) {
+        frontLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        frontRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+
         frontRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         backLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         frontLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
@@ -379,18 +433,129 @@ public class Left2 extends LinearOpMode {
         frontLeft.setPower(power);
         backRight.setPower(power);
         backLeft.setPower(power);
-        spin2(powers,times);
+        spin(powers,times);
         while (frontLeft.isBusy() && opModeIsActive()) {
 
         }
     }
+    public void moveandspinandcrane(double power, int position,double powers, int positions,double powerc, int timec) {
+        frontLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        frontRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+
+        frontRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        backLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        frontLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        backRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+
+        frontRight.setTargetPosition(-position);
+        frontLeft.setTargetPosition(-position);
+        backRight.setTargetPosition(-position);
+        backLeft.setTargetPosition(-position);
+
+        frontLeft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        frontRight.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        backLeft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        backRight.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+
+        frontRight.setPower(power);
+        frontLeft.setPower(power);
+        backRight.setPower(power);
+        backLeft.setPower(power);
+        spin(powers,positions);
+        crane(powerc,timec);
+        while (frontLeft.isBusy() && opModeIsActive()) {
+
+        }
+    }
+    public void moveandcrane(double power, int position,double powerc, int timec) {
+        frontLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        frontRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+
+        frontRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        backLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        frontLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        backRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+
+        frontRight.setTargetPosition(-position);
+        frontLeft.setTargetPosition(-position);
+        backRight.setTargetPosition(-position);
+        backLeft.setTargetPosition(-position);
+
+        frontLeft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        frontRight.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        backLeft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        backRight.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+
+        frontRight.setPower(power);
+        frontLeft.setPower(power);
+        backRight.setPower(power);
+        backLeft.setPower(power);
+        crane(powerc,timec);
+        while (frontLeft.isBusy() && opModeIsActive()) {
+
+        }
+    }
+    public void movewithdistance(double power,double distances) {
+        frontRight.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        frontLeft.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        backRight.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        backLeft.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        while(distance.getDistance(DistanceUnit.INCH)>=distances && opModeIsActive()){
+            //frontRight.setPower(-power);
+            //frontLeft.setPower(power);
+            //backRight.setPower(power);
+            //backLeft.setPower(-power-.2);
+            frontLeft.setPower(power);
+            frontRight.setPower(power);
+            backLeft.setPower(power);
+            backRight.setPower(power);
+
+        }
+        stopMotors();
+        frontRight.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        frontLeft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        backRight.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        backLeft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+    }
+
+    public void movewithdistance2(double power,double distances) {
+        frontRight.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        frontLeft.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        backRight.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        backLeft.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        while(distance2.getDistance(DistanceUnit.INCH)>=distances && opModeIsActive()){
+            //frontRight.setPower(-power);
+            //frontLeft.setPower(power);
+            //backRight.setPower(power);
+            //backLeft.setPower(-power-.2);
+            frontLeft.setPower(power);
+            frontRight.setPower(power);
+            backLeft.setPower(power);
+            backRight.setPower(power);
+
+        }
+        stopMotors();
+        frontRight.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        frontLeft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        backRight.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        backLeft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+    }
+
     public void craneinput(int time) {
         crane(1,time);
         intake(1,1000);
-        sleep(1500);
+        sleep(100);
     }
 
     public void strafeLeftwithcrane(double power, int position,double powerc, int timec)  {
+        frontLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        frontRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
         frontRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         backLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
@@ -407,39 +572,62 @@ public class Left2 extends LinearOpMode {
         backLeft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         backRight.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
-        frontRight.setPower(power);
-        frontLeft.setPower(power);
-        backRight.setPower(power);
-        backLeft.setPower(power);
+
+        frontLeft.setPower(1);
+        frontRight.setPower(-1);
+        backLeft.setPower(-.75);
+        backRight.setPower(1);
+        //frontRight.setPower(power);
+        //frontLeft.setPower(power);
+        //backRight.setPower(power);
+        //backLeft.setPower(power);
         crane(powerc,timec);
         while (frontLeft.isBusy() && opModeIsActive()) {
         }
 
     }
-    public void strafeLeftwithcrane2(double power, int distances,double powerc, int timec) {
+    public void strafeLeftwithcraneandspin(double power, int position,double powerc, int timec, int powers, int positions)  {
+        frontLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        frontRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
-        distancesnap=distance.getDistance(DistanceUnit.INCH);
+        frontRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        backLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        frontLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        backRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
 
-        frontRight.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-        frontLeft.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-        backRight.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-        backLeft.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-        crane(powerc, timec);
-        while(distance.getDistance(DistanceUnit.INCH)<distances && opModeIsActive()){
-            frontRight.setPower(-power);
-            frontLeft.setPower(power);
-            backRight.setPower(power);
-            backLeft.setPower(-power);
+        frontRight.setTargetPosition(-position);
+        frontLeft.setTargetPosition(position);
+        backRight.setTargetPosition(position);
+        backLeft.setTargetPosition(-position);
 
+        frontLeft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        frontRight.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        backLeft.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        backRight.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+
+
+        frontLeft.setPower(1);
+        frontRight.setPower(-1);
+        backLeft.setPower(-.75);
+        backRight.setPower(1);
+        //frontRight.setPower(power);
+        //frontLeft.setPower(power);
+        //backRight.setPower(power);
+        //backLeft.setPower(power);
+        spin(powers,positions);
+        crane(powerc,timec);
+        while (frontLeft.isBusy() && opModeIsActive()) {
         }
-        stopMotors();
-        frontRight.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        frontLeft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        backRight.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        backLeft.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
     }
     public void strafeRightwithcrane(double power, int position,double powerc, int timec)  {
+        frontLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        frontRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+
         frontRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         backLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         frontLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
@@ -465,6 +653,10 @@ public class Left2 extends LinearOpMode {
         }
     }
     public void strafeLeft(double power, int position)  {
+        frontLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        frontRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
         frontRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         backLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
@@ -485,12 +677,16 @@ public class Left2 extends LinearOpMode {
         frontLeft.setPower(power);
         backRight.setPower(power);
         backLeft.setPower(power);
-        crane(-1,2100);
         while (frontLeft.isBusy() && opModeIsActive()) {
         }
 
     }
     public void strafeRight(double power, int position)  {
+        frontLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        frontRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        backRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+
         frontRight.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         backLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         frontLeft.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
@@ -527,6 +723,7 @@ public class Left2 extends LinearOpMode {
     public void spin(double power, int position) {
         telemetry.addData("spin", Spin.getCurrentPosition());
         telemetry.update();
+        Spin.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
         Spin.setTargetPosition(position);
         Spin.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         Spin.setPower(power);
